+++ src/cli/commands/MenuCommand.ts
@@ -8,6 +8,8 @@ import path from "node:path";
 import { AeroflyFlightFormatter } from "../../core/formatter/AeroflyFlightFormatter.js";
 import { ExportFileAeroflyMainMcfExport } from "../../core/converter/ExportFileAeroflyMainMcfConverter.js";
 import { ExportFileAeroflyCustomMissionsTmcConverter } from "../../core/converter/ExportFileAeroflyCustomMissionsTmcConverter.js";
+import { ExportFileGeoJsonConverter } from "../../core/converter/ExportFileGeoJsonConverter.js";
+import { ExportFileKmlConverter } from "../../core/converter/ExportFileKmlConverter.js";
 
 export type MenuCommandMethod = Exclude<keyof MenuCommand, "controller" | "showMenuTitle" | "name" | "execute">;
 
@@ -129,18 +131,19 @@ export class MenuCommand extends ControllerCommand {
 
         const liveries = this.controller.aircraftService.getAircraft(aeroflyCodeAircraft)?.liveries ?? [];
 
-        const aeroflyCodeLivery = liveries
-            ? await select({
-                  message: "Aircraft livery",
-                  default: this.controller.getLivery(),
-                  choices: liveries
-                      .map((livery) => ({
-                          name: livery.name,
-                          value: livery.aeroflyCode !== "default" ? livery.aeroflyCode : "",
-                      }))
-                      .sort((a, b) => a.name.localeCompare(b.name)),
-              })
-            : "";
+        const aeroflyCodeLivery =
+            liveries.length > 1
+                ? await select({
+                      message: "Aircraft livery",
+                      default: this.controller.getLivery(),
+                      choices: liveries
+                          .map((livery) => ({
+                              name: livery.name,
+                              value: livery.aeroflyCode !== "default" ? livery.aeroflyCode : "",
+                          }))
+                          .sort((a, b) => a.name.localeCompare(b.name)),
+                  })
+                : "";
 
         this.controller.setAircraft(aeroflyCodeAircraft, aeroflyCodeLivery);
 
@@ -151,7 +154,7 @@ export class MenuCommand extends ControllerCommand {
         CliFormatter.showMenuTitle(["Fuel & Payload"]);
         const fuel = this.controller.getMaxFuel()
             ? await number({
-                  message: `Fuel (kg) - max ${this.controller.getMaxFuel()} kg`,
+                  message: `Fuel (kg) - max ${this.controller.getMaxFuel().toFixed()} kg`,
                   default: this.controller.getFuel(),
                   min: 0,
                   max: this.controller.getMaxFuel(),
@@ -166,7 +169,7 @@ export class MenuCommand extends ControllerCommand {
         const payload =
             maxPayload > 0
                 ? await number({
-                      message: `Payload (kg) - max ${maxPayload} kg`,
+                      message: `Payload (kg) - max ${maxPayload.toFixed()} kg`,
                       default: this.controller.getPayload(),
                       min: 0,
                       max: maxPayload,
@@ -277,6 +280,14 @@ export class MenuCommand extends ControllerCommand {
                 {
                     name: "Aerofly TMC custom user missions file",
                     value: ExportFileAeroflyCustomMissionsTmcConverter.fileExtension,
+                },
+                {
+                    name: "GeoJSON file",
+                    value: ExportFileGeoJsonConverter.fileExtension,
+                },
+                {
+                    name: "Keyhole Markup Language (KML) file",
+                    value: ExportFileKmlConverter.fileExtension,
                 },
             ],
         });