+++ src/core/api/AviationWeatherApi.ts
@@ -31,6 +31,95 @@ export interface AviationWeatherApiMetar {
     clouds: AviationWeatherApiCloud[];
 }
 
+type AviationWeatherApiRunwaySurface = "A" | "C" | "G" | "W" | "T" | "H";
+
+/**
+ * @see https://aviationweather.gov/data/api/#/Data/dataAirport
+ */
+export interface AviationWeatherApiRunway {
+    id: string;
+    dimension: string;
+    surface: AviationWeatherApiRunwaySurface;
+    alignment: string;
+}
+
+/**
+ * @see https://aviationweather.gov/data/api/#/Data/dataAirport
+ */
+export interface AviationWeatherApiFrequency {
+    type: string;
+    freq?: number;
+}
+
+/**
+ * @see https://aviationweather.gov/data/api/#/Data/dataAirport
+ */
+export interface AviationWeatherApiAirport {
+    icaoId: string;
+    name: string;
+    type: "ARP" | "HEL";
+    lat: number;
+    lon: number;
+
+    /**
+     * meters MSL
+     */
+    elev: number;
+    magdec: string;
+    rwyNum: string;
+    tower: "T" | "-" | null;
+    beacon: "B" | "-" | null;
+    runways: AviationWeatherApiRunway[];
+    freqs: AviationWeatherApiFrequency[] | string;
+}
+
+export type AviationWeatherApiNavaidType = "VORTAC" | "VOR/DME" | "TACAN" | "NDB" | "VOR";
+
+/**
+ * @see https://aviationweather.gov/data/api/#/Data/dataNavaid
+ */
+export interface AviationWeatherApiNavaidRaw {
+    id: string;
+    type: AviationWeatherApiNavaidType;
+    name: string;
+    lat: number | string;
+    lon: number | string;
+    elev: number | string;
+    freq: number | string;
+    mag_dec: string;
+}
+
+export interface AviationWeatherApiNavaid {
+    id: string;
+    type: AviationWeatherApiNavaidType;
+    name: string;
+    lat: number;
+    lon: number;
+
+    /**
+     * meters MSL
+     */
+    elev: number;
+
+    /**
+     * in kHz for NDB, MHz for VOR/TACAN
+     * @see https://aviationweather.gov/data/api/#/Data/dataNavaid
+     */
+    freq: number;
+
+    /**
+     * with "+" to the east and "-" to the west. Substracted from a true heading this will give the magnetic heading.
+     */
+    mag_dec: number;
+}
+
+export interface AviationWeatherApiFix {
+    id: string;
+    type: "I" | "L" | "H" | "S" | "D" | "-" | string;
+    lat: number;
+    lon: number;
+}
+
 export class AviationWeatherApi {
     static async fetchMetar(ids: string[], date: Date | null = null): Promise<AviationWeatherApiMetar[]> {
         return AviationWeatherApi.doRequest<AviationWeatherApiMetar[]>(
@@ -71,6 +160,73 @@ export class AviationWeatherApi {
         );
     }
 
+    static async fetchAirports(ids: string[]): Promise<AviationWeatherApiAirport[]> {
+        return AviationWeatherApi.doRequest(
+            "/api/data/airport",
+            new URLSearchParams({
+                ids: ids.join(","),
+                // bbox: AviationWeatherApi.buildBbox(position, distance).join(","),
+                format: "json",
+            }),
+        );
+    }
+
+    static async fetchNavaids(ids: string[]): Promise<AviationWeatherApiNavaid[]> {
+        return AviationWeatherApi.doRequest<AviationWeatherApiNavaidRaw[]>(
+            "/api/data/navaid",
+            new URLSearchParams({
+                ids: ids.join(","),
+                format: "json",
+                // bbox: AviationWeatherApi.buildBbox(position, distance).join(","),
+            }),
+        ).then((data) => {
+            return AviationWeatherApi.normalizeNavAid(data);
+        });
+    }
+
+    static async fetchFix(ids: string[]): Promise<AviationWeatherApiFix[]> {
+        return AviationWeatherApi.doRequest<AviationWeatherApiFix[]>(
+            "/api/data/fix",
+            new URLSearchParams({
+                ids: ids.join(","),
+                format: "json",
+                // bbox: AviationWeatherApi.buildBbox(position, distance).join(","),
+            }),
+        );
+    }
+
+    /**
+     * @param position center of search area
+     * @param distance in meters, default 1000
+     * @see https://aviationweather.gov/data/api/#/Data/dataNavaid
+     * @returns {Promise<AviationWeatherApiNavaid[]>}
+     */
+    static async fetchNavaidsByPosition(position: Point, distance: number = 1000): Promise<AviationWeatherApiNavaid[]> {
+        return AviationWeatherApi.doRequest<AviationWeatherApiNavaidRaw[]>(
+            "/api/data/navaid",
+            new URLSearchParams({
+                // ids: ids.join(","),
+                format: "json",
+                bbox: AviationWeatherApi.buildBbox(position, distance).join(","),
+            }),
+        ).then((data) => {
+            return AviationWeatherApi.normalizeNavAid(data);
+        });
+    }
+
+    private static normalizeNavAid(data: AviationWeatherApiNavaidRaw[]): AviationWeatherApiNavaid[] {
+        return data.map((navaid: AviationWeatherApiNavaidRaw): AviationWeatherApiNavaid => {
+            return {
+                ...navaid,
+                lat: Number(navaid.lat),
+                lon: Number(navaid.lon),
+                elev: Number(navaid.elev),
+                freq: Number(navaid.freq),
+                mag_dec: magDecConverter(navaid.mag_dec),
+            };
+        });
+    }
+
     static async doRequest<T>(route: string, query: URLSearchParams): Promise<T> {
         const url = new URL(route + "?" + query, "https://aviationweather.gov");
         //console.log(url);
@@ -96,6 +252,131 @@ export class AviationWeatherApi {
         const southEast = position.getPointBy(new Vector(distance, 225));
         const northWest = position.getPointBy(new Vector(distance, 45));
         return [southEast.latitude, southEast.longitude, northWest.latitude, northWest.longitude];
+    }
+}
+
+/**
+ * @returns {number} with "+" to the east and "-" to the west. Substracted from a true heading this will give the magnetic heading.
+ */
+export const magDecConverter = (magdec: string): number => {
+    let magDec = 0;
+    const magdecMatch = magdec.match(/^(\d+)(E|W)$/);
+    if (magdecMatch) {
+        magDec = Number(magdecMatch[1]);
+        if (magdecMatch[2] === "W") {
+            magDec *= -1;
+        }
+    }
+    return magDec;
+};
+
+export class AviationWeatherNormalizedAirport {
+    icaoId: string;
+    name: string;
+    type: "ARP" | "HEL";
+    lat: number;
+    lon: number;
+
+    /**
+     * meters MSL
+     */
+    elev: number;
+    magdec: number;
+    rwyNum: number;
+    tower: boolean;
+    beacon: boolean;
+    runways: AviationWeatherNormalizedRunway[];
+    freqs: AviationWeatherApiFrequency[];
+
+    /**
+     * @param {AviationWeatherApiAirport} apiData
+     */
+    constructor({
+        icaoId,
+        name,
+        type,
+        lat,
+        lon,
+        elev,
+        magdec,
+        rwyNum,
+        tower,
+        beacon,
+        runways,
+        freqs,
+    }: AviationWeatherApiAirport) {
+        this.icaoId = icaoId;
+
+        this.name = name
+            .replace(/_/g, " ")
+            .trim()
+            .replace(/\bINTL\b/g, "INTERNATIONAL")
+            .replace(/\bRGNL\b/g, "REGIONAL")
+            .replace(/\bFLD\b/g, "FIELD")
+            .replace(/(\/)/g, " $1 ")
+            .toLowerCase()
+            .replace(/(^|\s)[a-z]/g, (char) => {
+                return char.toUpperCase();
+            });
+
+        this.type = type;
+        this.lat = lat;
+        this.lon = lon;
+        this.elev = elev;
+        this.magdec = magDecConverter(magdec);
+        this.rwyNum = Number(rwyNum);
+        this.tower = tower === "T";
+        this.beacon = beacon === "B";
+        this.runways = runways.map((r) => {
+            return new AviationWeatherNormalizedRunway(r);
+        });
+
+        this.freqs =
+            typeof freqs !== "string"
+                ? freqs
+                : freqs.split(";").map(
+                      /**
+                       * @param {string} f
+                       * @returns {AviationWeatherApiFrequency}
+                       */
+                      (f: string): AviationWeatherApiFrequency => {
+                          const parts = f.split(",");
+                          return {
+                              type: parts[0],
+                              freq: parts[1] ? Number(parts[1]) : undefined,
+                          };
+                      },
+                  );
+    }
+}
+
+export class AviationWeatherNormalizedRunway {
+    id: [string, string];
+
+    /**
+     * length, width in ft
+     */
+    dimension: [number, number];
+    surface: string;
+    alignment: number | null;
+
+    constructor({ id, dimension, surface, alignment }: AviationWeatherApiRunway) {
+        /**
+         * @type {[string,string]} both directions
+         */
+        this.id = ["", ""];
+        id.split("/").forEach((i, index) => {
+            this.id[index] = i;
+        });
+        this.dimension = [0, 0];
+        dimension
+            .split("x")
+            .map((x) => Number(x))
+            .forEach((d, index) => {
+                this.dimension[index] = d;
+            });
+        this.surface = surface;
+        this.alignment = alignment !== "-" ? Number(alignment) : null;
     }
 }