+++ src/core/io/ExportFileWriter.ts
@@ -3,6 +3,8 @@ import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
 import { ExportFileAeroflyMainMcfExport } from "../converter/ExportFileAeroflyMainMcfConverter.js";
 import { ExportFileConverter } from "../converter/ExportFileConverter.js";
 import { ExportFileAeroflyCustomMissionsTmcConverter } from "../converter/ExportFileAeroflyCustomMissionsTmcConverter.js";
+import { ExportFileGeoJsonConverter } from "../converter/ExportFileGeoJsonConverter.js";
+import { ExportFileKmlConverter } from "../converter/ExportFileKmlConverter.js";
 
 /**
  * Writes a file from an `AeroflyFlight` class instance to an
@@ -22,15 +24,21 @@ export class ExportFileWriter {
             throw new Error(`Could not determine file type for "${filename}"`);
         }
 
-        const registry: Record<string, new () => ExportFileConverter> = {
-            [ExportFileAeroflyMainMcfExport.fileExtension]: ExportFileAeroflyMainMcfExport,
-            [ExportFileAeroflyCustomMissionsTmcConverter.fileExtension]: ExportFileAeroflyCustomMissionsTmcConverter,
-        };
+        const registry = this.getRegistry();
 
         const converter = registry[fileSuffix];
         if (!converter) {
             throw new Error(`Unsupported file type: ${fileSuffix}`);
         }
         return converter;
+    }
+
+    static getRegistry(): Record<string, (new () => ExportFileConverter) | undefined> {
+        return {
+            [ExportFileAeroflyMainMcfExport.fileExtension]: ExportFileAeroflyMainMcfExport,
+            [ExportFileAeroflyCustomMissionsTmcConverter.fileExtension]: ExportFileAeroflyCustomMissionsTmcConverter,
+            [ExportFileGeoJsonConverter.fileExtension]: ExportFileGeoJsonConverter,
+            [ExportFileKmlConverter.fileExtension]: ExportFileKmlConverter,
+        };
     }
 }