+++ types/core/api/AviationWeatherApi.test.d.ts
@@ -1,0 +1,2 @@
+export {};
+//# sourceMappingURL=AviationWeatherApi.test.d.ts.map