+++ types/core/converter/ExportFileAeroflyMainMcfConverter.d.ts
@@ -1,6 +1,6 @@
 import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
 import { ExportFileConverter } from "./ExportFileConverter.js";
-export declare class ExportFileAeroflyMainMcfExport implements ExportFileConverter {
+export declare class ExportFileAeroflyMainMcfExport extends ExportFileConverter {
     static readonly fileExtension = "mcf";
     convert(flightplan: AeroflyFlight): string;
 }