+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Removed community-developed airports from base airport set
 - Improved creation of flight plans
+- Added KML and GeoJSON export
 
 ## [1.4.3]